requirements.txt
beautifulsoup4==4.11.1
lxml==4.9.1
html5lib==1.1
urllib3==1.26.12
python-dotenv==1.0.1
requests==2.31.0
fastapi==0.110.1
gunicorn==21.2.0
uvicorn==0.29.0
langchain==0.1.20
langchain-openai==0.1.6
imbox==0.9.8
